支持发送上传图片且默认原图

+++ Ore-Base/src/main/kotlin/moe/ore/helper/thread/ThreadManager.kt
@@ -24,7 +24,7 @@ class ThreadManager private constructor(val uin: Long = 0) {
         corePoolSize,
         (maxCachePoolSize + corePoolSize),
         // 线程池最大线程数
-        keepAliveTime.toLong(),
+        keepAliveTime,
         TimeUnit.MILLISECONDS,
         linkedBlockingQueue,
         handler
